* master: (7207 commits)
  Adding sink check on script element related tests in trusted types. (#46133)
  Add more tests to check mrow fallback when computing preferred logical widths (#47094)
  compute pressure: Make AddClient() return a pending receiver, not take a pending remote
  Harmonize manifest variables: remove unused OS, ARCH values
  Remove CanvasRedneringContext2D scrollPathIntoView() tests
  Part 6 - Skip style sharing revalidation for trivial styles.
  Part 3 - Add scope revalidation to style sharing.
  Don't fire popstate for history.pushState/replaceState intercepted by the navigate event
  Don't attempt to trim letter-spacing at start of line in textframes being used for SVG text.
  Text Fragments: Scroll the start container of a range into view instead of its common ancestor.
  [cookie wpt] make tests more reliable
  webnn: fix coreml handling for inputs with special characters
  webnn: add argmin/max on CoreML and add outputDataType
  Fenced frame: attempt to deflake ancestor-throttle test.
  [reading-flow] fix pseudo element reading traversal
  Multicol: Don't pass accumulated decorations to FinishFragmentation.
  Speculative fix for WebXR WPT already declared syntax error
  Add WPTs for 'allow-storage-access-by-user-activation' iframe sandbox (#47067)
  Expect short-value serialization in display-valid.html (#47029)
  MIME Sniffing: HTML no feed
  ...

+++ storage-access-api/helpers.js
@@ -29,8 +29,13 @@ async function CreateFrameHelper(setUpFrame, fetchTests) {
 // Create an iframe element with content loaded from `sourceURL`, append it to
 // the document, and optionally fetch tests. Returns the loaded frame, once
 // ready.
-function CreateFrame(sourceURL, fetchTests = false) {
+function CreateFrame(
+  sourceURL, fetchTests = false, frameSandboxAttribute = undefined) {
   return CreateFrameHelper((frame) => {
+    if (frameSandboxAttribute !== undefined) {
+      frame.sandbox = frameSandboxAttribute;
+    }
+
     frame.src = sourceURL;
     document.body.appendChild(frame);
   }, fetchTests);
@@ -38,8 +43,8 @@ function CreateFrame(sourceURL, fetchTests = false) {
 
 // Create a new iframe with content loaded from `sourceURL`, and fetches tests.
 // Returns the loaded frame, once ready.
-function RunTestsInIFrame(sourceURL) {
-  return CreateFrame(sourceURL, true);
+function RunTestsInIFrame(sourceURL, frameSandboxAttribute = undefined) {
+  return CreateFrame(sourceURL, true, frameSandboxAttribute);
 }
 
 function RunTestsInNestedIFrame(sourceURL) {
@@ -57,18 +62,18 @@ function RunTestsInNestedIFrame(sourceURL) {
   }, true);
 }
 
-function RunRequestStorageAccessInDetachedFrame() {
+function CreateDetachedFrame() {
   const frame = document.createElement('iframe');
   document.body.append(frame);
   const inner_doc = frame.contentDocument;
   frame.remove();
-  return inner_doc.requestStorageAccess();
+  return inner_doc;
 }
 
-function RunRequestStorageAccessViaDomParser() {
+function CreateDocumentViaDOMParser() {
   const parser = new DOMParser();
   const doc = parser.parseFromString('<html></html>', 'text/html');
-  return doc.requestStorageAccess();
+  return doc;
 }
 
 function RunCallbackWithGesture(callback) {
@@ -126,11 +131,11 @@ async function DeleteCookieInFrame(frame, name, params) {
 }
 
 // Tests whether the frame can write cookies via document.cookie. Note that this
-// overwrites, then deletes, cookies named "cookie" and "foo".
+// overwrites, then optionally deletes, cookies named "cookie" and "foo".
 //
 // This function requires the caller to have included
 // /cookies/resources/cookie-helper.sub.js.
-async function CanFrameWriteCookies(frame) {
+async function CanFrameWriteCookies(frame, keep_after_writing = false) {
   const cookie_suffix = "Secure;SameSite=None;Path=/";
   await DeleteCookieInFrame(frame, "cookie", cookie_suffix);
   await DeleteCookieInFrame(frame, "foo", cookie_suffix);
@@ -142,10 +147,65 @@ async function CanFrameWriteCookies(frame) {
   const can_write = cookieStringHasCookie("cookie", "monster", cookies) &&
       cookieStringHasCookie("foo", "bar", cookies);
 
-  await DeleteCookieInFrame(frame, "cookie", cookie_suffix);
-  await DeleteCookieInFrame(frame, "foo", cookie_suffix);
+  if (!keep_after_writing) {
+    await DeleteCookieInFrame(frame, "cookie", cookie_suffix);
+    await DeleteCookieInFrame(frame, "foo", cookie_suffix);
+  }
 
   return can_write;
+}
+
+// Sets a cookie in an unpartitioned context by creating a new frame
+// and requesting storage access in the frame.
+async function SetFirstPartyCookieAndUnsetStorageAccessPermission(origin) {
+  let frame = await CreateFrame(`${origin}/storage-access-api/resources/script-with-cookie-header.py?script=embedded_responder.js`);
+  await SetPermissionInFrame(frame, [{ name: 'storage-access' }, 'granted']);
+  await RequestStorageAccessInFrame(frame);
+  await SetDocumentCookieFromFrame(frame, `cookie=unpartitioned;Secure;SameSite=None;Path=/`);
+  await SetPermissionInFrame(frame, [{ name: 'storage-access' }, 'prompt']);
+}
+
+// Tests for the presence of the unpartitioned cookie set by SetFirstPartyCookieAndUnsetStorageAccessPermission
+// in both the `document.cookie` variable and same-origin subresource \
+// Request Headers in the given frame
+async function HasUnpartitionedCookie(frame) {
+  let frameDocumentCookie = await GetJSCookiesFromFrame(frame);
+  let jsAccess = cookieStringHasCookie("cookie", "unpartitioned", frameDocumentCookie);
+  const httpCookie = await FetchSubresourceCookiesFromFrame(frame, "");
+  let httpAccess = cookieStringHasCookie("cookie", "unpartitioned", httpCookie);
+  assert_equals(jsAccess, httpAccess, "HTTP and Javascript cookies must be in sync");
+  return jsAccess && httpAccess;
+}
+
+// Tests whether the current frame can read and write cookies via HTTP headers.
+// This deletes, writes, reads, then deletes a cookie named "cookie".
+async function CanAccessCookiesViaHTTP() {
+  // We avoid reusing SetFirstPartyCookieAndUnsetStorageAccessPermission here, since that bypasses the
+  // cookie-accessibility settings that we want to check here.
+  await fetch(`${window.location.origin}/storage-access-api/resources/set-cookie-header.py?cookie=1;path=/;SameSite=None;Secure`);
+  const http_cookies = await fetch(`${window.location.origin}/storage-access-api/resources/echo-cookie-header.py`)
+      .then((resp) => resp.text());
+  const can_access = cookieStringHasCookie("cookie", "1", http_cookies);
+
+  erase_cookie_from_js("cookie", "SameSite=None;Secure;Path=/");
+
+  return can_access;
+}
+
+// Tests whether the current frame can read and write cookies via
+// document.cookie. This deletes, writes, reads, then deletes a cookie named
+// "cookie".
+function CanAccessCookiesViaJS() {
+  erase_cookie_from_js("cookie", "SameSite=None;Secure;Path=/");
+  assert_false(cookieStringHasCookie("cookie", "1", document.cookie));
+
+  document.cookie = "cookie=1;SameSite=None;Secure;Path=/";
+  const can_access = cookieStringHasCookie("cookie", "1", document.cookie);
+
+  erase_cookie_from_js("cookie", "SameSite=None;Secure;Path=/");
+  assert_false(cookieStringHasCookie("cookie", "1", document.cookie));
+
+  return can_access;
 }
 
 // Reads cookies via the `httpCookies` variable in the given frame.
@@ -196,10 +256,25 @@ function FrameInitiatedNavigation(frame, url) {
   return load;
 }
 
-// Makes a subresource request to the provided host in the given frame, and returns the cookies in the response.
-function FetchFromFrame(frame, host) {
+// Makes a subresource request to the provided host in the given frame, and
+// returns the cookies that were included in the request.
+function FetchSubresourceCookiesFromFrame(frame, host) {
+  return FetchFromFrame(frame, `${host}/storage-access-api/resources/echo-cookie-header.py`);
+}
+
+// Makes a subresource request to the provided host in the given frame, and
+// returns the response.
+function FetchFromFrame(frame, url) {
   return PostMessageAndAwaitReply(
-    { command: "subresource cookies", host }, frame.contentWindow);
+    { command: "cors fetch", url }, frame.contentWindow);
+}
+
+// Makes a subresource request to the provided host in the given frame with
+// the mode set to 'no-cors'
+function NoCorsSubresourceCookiesFromFrame(frame, host) {
+  const url = `${host}/storage-access-api/resources/echo-cookie-header.py`;
+  return PostMessageAndAwaitReply(
+    { command: "no-cors fetch", url }, frame.contentWindow);
 }
 
 // Tries to set storage access policy, ignoring any errors.
@@ -215,4 +290,29 @@ async function MaybeSetStorageAccess(origin, embedding_origin, value) {
     // by default. If this failed without default blocking we'll notice it later
     // in the test.
   }
+}
+
+
+// Navigate the inner iframe using the given frame.
+function NavigateChild(frame, url) {
+  return PostMessageAndAwaitReply(
+    { command: "navigate_child", url }, frame.contentWindow);
+}
+
+// Starts a dedicated worker in the given frame.
+function StartDedicatedWorker(frame) {
+  return PostMessageAndAwaitReply(
+    { command: "start_dedicated_worker" }, frame.contentWindow);
+}
+
+// Sends a message to the dedicated worker in the given frame.
+function MessageWorker(frame, message = {}) {
+  return PostMessageAndAwaitReply(
+    { command: "message_worker", message }, frame.contentWindow);
+}
+// Opens a WebSocket connection to origin from within frame, and
+// returns the cookie header that was sent during the handshake.
+function ReadCookiesFromWebSocketConnection(frame, origin) {
+  return PostMessageAndAwaitReply(
+   { command: "get_cookie_via_websocket", origin}, frame.contentWindow);
 }
